Add Cirrus CI support for Linux/macOS/Windows
This change (which I'm pretty excited about) adds continuous integration
backed by Cirrus CI to the project. Linux, macOS, and Windows platforms
are supported for builds and tests across latest stable and nightly Rust
versions and a dedicated job for "checking" (i.e. Formatting and
linting) is provided to separate the "does it work" from "is the
formatting consistent".

This project is using the [cargo-make] Cargo plugin for consistent,
cross platform workflows and the plugin itself is cached between CI runs
for greater execution speed.

Note that while the YAML description is rather long and nontrivial, it
is extremely effective and can be used/adapted/copied into future
projects with all the benefit. In short: it was very well worth the
effort getting this to work.

[cargo-make]: https://sagiegurari.github.io/cargo-make/

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>