[mtoc] Expose `Writer` to public API.
This change exposes one previously internal type to the public API and
removes one internal type. The `Writer` struct now performs the duties
of `WriterFormatBuilder` and `Writer` by making the `headers` field an
`Option`. The main reason for the additional `Writer` type is to hide
the need to declare the generic `Format` type in the main
`WriterBuilder` type for most invocations. The prior public API is
preserved but now consumers of this crate can get a direct reference to
the `Writer` produced from a `WriterBuilder`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>