[ci] Use stable Rust to build cargo-make
There is a distinction between building the project source with a target
version of Rust and using the cargo-make tool to drive that. In other
words the version of Rust that build cargo-make should probably be
independent of the version of Rust that build the project.

In this case, I found that cargo-make wouldn't build on Rust 1.34.0
which was a minimum Rust version check in this project.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>